separate levenshteinDistance function to its own file/module
