+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { configure } from './configure';
 export { generateMock } from './generateMock';
+export { levenshteinDistance } from './levenshteinDistance';
